# providers/keycloak/src/airflow/providers/keycloak/auth_manager/api_fastapi/datamodels/security.py
# Licensed to the Apache Software Foundation (ASF) under one
# or more contributor license agreements.  See the NOTICE file
# distributed with this work for additional information
# regarding copyright ownership.  The ASF licenses this file
# to you under the Apache License, Version 2.0 (the
# "License"); you may not use this file except in compliance
# with the License.  You may obtain a copy of the License at
#
#   http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing,
# software distributed under the License is distributed on an
# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
# KIND, either express or implied.  See the License for the
# specific language governing permissions and limitations
# under the License.
from typing import Optional

from airflow.api_fastapi.core_api.base import StrictBaseModel
from pydantic import BaseModel, EmailStr, Field


class UserResponse(BaseModel):
    userid: str
    username: str
    email: EmailStr
    first_name: str
    last_name: str
    enabled: bool
    email_verified: bool
    roles: list[str]
    created_timestamp: Optional[int] = None

class UserRequest(StrictBaseModel):
    username: str = Field()
    email: EmailStr = Field()
    first_name: str = Field()
    last_name: str = Field()
    enabled: bool = True
    email_verified: bool = False
    role: str = Field()
    password: str = Field()


class UserUpdateRequest(StrictBaseModel):
    username: Optional[str] = Field(default=None)
    email: Optional[EmailStr] = Field(default=None)
    firstName: Optional[str] = Field(default=None)
    lastName: Optional[str] = Field(default=None)
    enabled: Optional[bool] = Field(default=None)
    emailVerified: Optional[bool] = Field(default=None)
    role: Optional[str] = Field(default=None)
    password: Optional[str] = Field(default=None)

# class KeycloakRoleRequest(BaseModel):
#     name: str
#     description: Optional[str] = None
#     composite: bool = False
#     clientRole: bool = False
#     containerId: Optional[str] = None
#
# class KeycloakGroupRequest(BaseModel):
#     name: str
#     path: Optional[str] = None
#     attributes: Optional[Dict[str, List[str]]] = None
#     realmRoles: Optional[List[str]] = None
#     clientRoles: Optional[Dict[str, List[str]]] = None